# #451 optimize evaluation performance