… Jint/Native/String/StringPrototype.cs
… Jint/Pooling/JsValueArrayPool.cs
… Jint/Runtime/Arguments.cs
… Jint/Runtime/Debugger/DebugHandler.cs
… Jint/Runtime/Environments/DeclarativeEnvironmentRecord.cs
… Jint/Runtime/ExpressionIntepreter.cs
… Jint/Runtime/Interop/DefaultTypeConverter.cs
… Jint/StrictModeScope.cs