* make StrictModeScope a struct
* prioritize statement branching
* reuse JsValue array instances
* check string length before equality (can save some 20% if less likely to hit)


## Esprima.Benchmark.DromaeoBenchmark

| **Diff**|Method|FileName|Mean|Gen 0|Gen 1|Gen 2|Allocated|
|------- |-------|-------|-------:|-------:|-------:|-------:|-------:|
| Old |Run|dromaeo-3d-cube|103.35 ms|1312.5000|250.0000|-|7107.59 KB|
| **New** |	|	| **101.68 ms (-2%)** | **1250.0000 (-5%)** | **312.5000 (+25%)** | **-** | **6812.54 KB (-4%)** |
| Old |Run|dromaeo-core-eval|24.70 ms|62.5000|-|-|298.95 KB|
| **New** |	|	| **23.89 ms (-3%)** | **62.5000 (0%)** | **-** | **-** | **298.55 KB (0%)** |
| Old |Run|dromaeo-object-array|259.58 ms|39500.0000|2000.0000|1000.0000|166027.24 KB|
| **New** |	|	| **255.24 ms (-2%)** | **39125.0000 (-1%)** | **2000.0000 (0%)** | **1000.0000 (0%)** | **164345.05 KB (-1%)** |
| Old |Run|dromaeo-object-regexp|1,142.73 ms|57750.0000|34750.0000|22125.0000|433170.74 KB|
| **New** |	|	| **1,158.63 ms (+1%)** | **56437.5000 (-2%)** | **34687.5000 (0%)** | **21812.5000 (-1%)** | **430615.03 KB (-1%)** |
| Old |Run|dromaeo-object-string|1,359.54 ms|220812.5000|178437.5000|175812.5000|1392803.07 KB|
| **New** |	|	| **1,360.40 ms (0%)** | **218375.0000 (-1%)** | **177000.0000 (-1%)** | **174187.5000 (-1%)** | **1390250.42 KB (0%)** |
| Old |Run|dromaeo-string-base64|234.61 ms|6187.5000|187.5000|-|26458.12 KB|
| **New** |	|	| **229.69 ms (-2%)** | **5437.5000 (-12%)** | **437.5000 (+133%)** | **-** | **23300.13 KB (-12%)** |


## Jint.Benchmark.ArrayBenchmark

| **Diff**|Method|N|Mean|Gen 0|Allocated|
|------- |-------|-------|-------:|-------:|-------:|
| Old |Slice|100|671.5 us|145.5078|598.44 KB|
| **New** |	|	| **666.6 us (-1%)** | **145.5078 (0%)** | **596.09 KB (0%)** |
| Old |Concat|100|797.6 us|162.1094|666.41 KB|
| **New** |	|	| **787.3 us (-1%)** | **160.1563 (-1%)** | **657.81 KB (-1%)** |
| Old |Unshift|100|30,480.2 us|3031.2500|12525.78 KB|
| **New** |	|	| **30,568.4 us (0%)** | **2968.7500 (-2%)** | **12267.19 KB (-2%)** |
| Old |Push|100|17,942.4 us|593.7500|2518.75 KB|
| **New** |	|	| **17,673.1 us (-2%)** | **531.2500 (-11%)** | **2260.16 KB (-10%)** |
| Old |Index|100|16,595.0 us|343.7500|1515.63 KB|
| **New** |	|	| **16,299.8 us (-2%)** | **343.7500 (0%)** | **1510.16 KB (0%)** |
| Old |Map|100|4,150.7 us|796.8750|3286.72 KB|
| **New** |	|	| **4,050.0 us (-2%)** | **750.0000 (-6%)** | **3083.59 KB (-6%)** |
| Old |Apply|100|933.3 us|176.7578|727.34 KB|
| **New** |	|	| **941.8 us (+1%)** | **175.7813 (-1%)** | **721.09 KB (-1%)** |
| Old |JsonStringifyParse|100|5,289.6 us|1242.1875|5111.72 KB|
| **New** |	|	| **5,127.9 us (-3%)** | **1242.1875 (0%)** | **5103.13 KB (0%)** |


## Jint.Benchmark.ArrayStressBenchmark

| **Diff**|Method|N|Mean|Gen 0|Gen 1|Allocated|
|------- |-------|-------|-------:|-------:|-------:|-------:|
| Old |Jint|20|1.646 s|60000.0000|5000.0000|249.81 MB|
| **New** |	|	| **1.510 s (-8%)** | **59937.5000 (0%)** | **1062.5000 (-79%)** | **241.6 MB (-3%)** |


## Jint.Benchmark.UncacheableExpressionsBenchmark

| **Diff**|Method|N|Mean|Gen 0|Gen 1|Allocated|
|------- |-------|-------|-------:|-------:|-------:|-------:|
| Old |Benchmark|500|461.2 ms|52812.5000|62.5000|211.35 MB|
| **New** |	|	| **464.0 ms (+1%)** | **50375.0000 (-5%)** | **375.0000 (+500%)** | **201.85 MB (-4%)** |